Switched to using published for date of atom posts
This should be made more robust in the future.  For example, one could
check both and use the one that's present, or even expose both on the
common `Item` struct.

This is a temporary fix, mostly for YouTube, which sets updated quite
often, likely every time the description is updated.